+++ graphql_api/mutation/record_label_association.py
@@ -31,7 +31,7 @@ class CreateClassificationAssociation(graphene.Mutation):
         labeling_task_id: str,
         as_gold_star: Optional[bool] = None,
     ):
-        auth.check_is_demo(info)
+        auth.check_demo_access(info)
         auth.check_project_access(info, project_id)
         user = auth.get_user_by_info(info)
         record = manager.create_classification_label(
@@ -78,7 +78,7 @@ class CreateExtractionAssociation(graphene.Mutation):
         value: str,
         as_gold_star: Optional[bool] = None,
     ):
-        auth.check_is_demo(info)
+        auth.check_demo_access(info)
         auth.check_project_access(info, project_id)
         user = auth.get_user_by_info(info)
         record = manager.create_extraction_label(
@@ -121,7 +121,7 @@ class SetGoldStarAnnotationForTask(graphene.Mutation):
         labeling_task_id: str,
         gold_user_id: str,
     ):
-        auth.check_is_demo(info)
+        auth.check_demo_access(info)
         auth.check_project_access(info, project_id)
         user = auth.get_user_by_info(info)
         task_type = manager.create_gold_star_association(
@@ -150,7 +150,7 @@ class DeleteRecordLabelAssociationByIds(graphene.Mutation):
     ok = graphene.Boolean()
 
     def mutate(self, info, project_id: str, record_id: str, association_ids: List[str]):
-        auth.check_is_demo(info)
+        auth.check_demo_access(info)
         auth.check_project_access(info, project_id)
         manager.delete_record_label_association(project_id, record_id, association_ids)
         notification.send_organization_update(project_id, f"rla_deleted:{record_id}")
@@ -167,7 +167,7 @@ class DeleteGoldStarAssociationForTask(graphene.Mutation):
     ok = graphene.Boolean()
 
     def mutate(self, info, project_id: str, record_id: str, labeling_task_id: str):
-        auth.check_is_demo(info)
+        auth.check_demo_access(info)
         auth.check_project_access(info, project_id)
         user = auth.get_user_by_info(info)
         manager.delete_gold_star_association(
@@ -186,7 +186,7 @@ class UpdateRlaIsValidManual(graphene.Mutation):
     ok = graphene.Boolean()
 
     def mutate(self, info, execute: Optional[bool] = False):
-        auth.check_is_demo(info)
+        auth.check_demo_access(info)
         if execute:
             manager.update_is_valid_manual_label_for_all()
